Patient management interface completed; code refactored
Patient management interface completed, fixed missing functions for hospital, refactoeed code into several files
