+++ MIAN.CPP
@@ -1,0 +1,18 @@
+void main()
+{
+	clrscr();
+    
+    /*/////Administrator object creator////////////////
+	address yay = {0, 0, 0, 0, 0};
+	employee x("Administrator", 3, Date(), yay, "", 0, Time(), Time(), "admin", "password");
+	hospital::write_employee(&x);
+    /////////////////////////////////////////////*/
+	
+    interface::log_this("Program initiated\n-----------------------------------\n");
+
+	interface::init();
+
+	interface::log_this("Program terminated\n-----------------------------------\n");
+
+	getch();
+}
